refactor seeding data to break it into different files
